requirements.txt
firebase_admin==6.2.0
Flask==2.2.2
Pillow==10.0.1
requests==2.28.1
